fix(metrics): implement component-defined metrics system
