+++ src/vmm/src/devices/virtio/pmem/metrics.rs
@@ -84,7 +84,7 @@ use std::sync::{Arc, RwLock};
 use serde::ser::SerializeMap;
 use serde::{Serialize, Serializer};
 
-use crate::logger::{IncMetric, LatencyAggregateMetrics, SharedIncMetric};
+use crate::metrics::{IncMetric, LatencyAggregateMetrics, SharedIncMetric};
 
 /// map of pmem drive id and metrics
 /// this should be protected by a lock before accessing.
