+++ src/vmm/src/devices/virtio/transport/mmio.rs
@@ -15,7 +15,8 @@ use super::{VirtioInterrupt, VirtioInterruptType};
 use crate::devices::virtio::device::VirtioDevice;
 use crate::devices::virtio::device_status;
 use crate::devices::virtio::queue::Queue;
-use crate::logger::{IncMetric, METRICS, error, warn};
+use crate::logger::{METRICS, error, warn};
+use crate::metrics::IncMetric;
 use crate::utils::byte_order;
 use crate::vstate::bus::BusDevice;
 use crate::vstate::interrupts::InterruptError;
