+++ src/vmm/src/logger/metrics.rs
@@ -64,12 +64,13 @@
 use std::fmt::Debug;
 use std::io::Write;
 use std::ops::Deref;
-use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{Mutex, OnceLock};
 
 use serde::{Serialize, Serializer};
 use utils::time::{ClockType, get_time_ns, get_time_us};
 
+// Import metric traits and implementations from the metrics module
+use crate::metrics::{LatencyAggregateMetrics, SharedIncMetric, SharedStoreMetric, StoreMetric};
 use super::FcLineWriter;
 use crate::devices::legacy;
 use crate::devices::virtio::balloon::metrics as balloon_metrics;
@@ -174,110 +175,6 @@ pub enum MetricsError {
     Serde(String),
     /// Failed to write metrics: {0}
     Write(std::io::Error),
-}
-
-/// Used for defining new types of metrics that act as a counter (i.e they are continuously updated
-/// by incrementing their value).
-pub trait IncMetric {
-    /// Adds `value` to the current counter.
-    fn add(&self, value: u64);
-    /// Increments by 1 unit the current counter.
-    fn inc(&self) {
-        self.add(1);
-    }
-    /// Returns current value of the counter.
-    fn count(&self) -> u64;
-
-    /// Returns diff of current and old value of the counter.
-    /// Mostly used in process of aggregating per device metrics.
-    fn fetch_diff(&self) -> u64;
-}
-
-/// Used for defining new types of metrics that do not need a counter and act as a persistent
-/// indicator.
-pub trait StoreMetric {
-    /// Returns current value of the counter.
-    fn fetch(&self) -> u64;
-    /// Stores `value` to the current counter.
-    fn store(&self, value: u64);
-}
-
-/// Representation of a metric that is expected to be incremented from more than one thread, so more
-/// synchronization is necessary.
-// It's currently used for vCPU metrics. An alternative here would be
-// to have one instance of every metric for each thread, and to
-// aggregate them when writing. However this probably overkill unless we have a lot of vCPUs
-// incrementing metrics very often. Still, it's there if we ever need it :-s
-// We will be keeping two values for each metric for being able to reset
-// counters on each metric.
-// 1st member - current value being updated
-// 2nd member - old value that gets the current value whenever metrics is flushed to disk
-#[derive(Debug, Default)]
-pub struct SharedIncMetric(AtomicU64, AtomicU64);
-impl SharedIncMetric {
-    /// Const default construction.
-    pub const fn new() -> Self {
-        Self(AtomicU64::new(0), AtomicU64::new(0))
-    }
-}
-
-/// Representation of a metric that is expected to hold a value that can be accessed
-/// from more than one thread, so more synchronization is necessary.
-#[derive(Debug, Default)]
-pub struct SharedStoreMetric(AtomicU64);
-impl SharedStoreMetric {
-    /// Const default construction.
-    pub const fn new() -> Self {
-        Self(AtomicU64::new(0))
-    }
-}
-
-impl IncMetric for SharedIncMetric {
-    // While the order specified for this operation is still Relaxed, the actual instruction will
-    // be an asm "LOCK; something" and thus atomic across multiple threads, simply because of the
-    // fetch_and_add (as opposed to "store(load() + 1)") implementation for atomics.
-    // TODO: would a stronger ordering make a difference here?
-    fn add(&self, value: u64) {
-        self.0.fetch_add(value, Ordering::Relaxed);
-    }
-
-    fn count(&self) -> u64 {
-        self.0.load(Ordering::Relaxed)
-    }
-    fn fetch_diff(&self) -> u64 {
-        self.0.load(Ordering::Relaxed) - self.1.load(Ordering::Relaxed)
-    }
-}
-
-impl StoreMetric for SharedStoreMetric {
-    fn fetch(&self) -> u64 {
-        self.0.load(Ordering::Relaxed)
-    }
-
-    fn store(&self, value: u64) {
-        self.0.store(value, Ordering::Relaxed);
-    }
-}
-
-impl Serialize for SharedIncMetric {
-    /// Reset counters of each metrics. Here we suppose that Serialize's goal is to help with the
-    /// flushing of metrics.
-    /// !!! Any print of the metrics will also reset them. Use with caution !!!
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        let snapshot = self.0.load(Ordering::Relaxed);
-        let res = serializer.serialize_u64(snapshot - self.1.load(Ordering::Relaxed));
-
-        if res.is_ok() {
-            self.1.store(snapshot, Ordering::Relaxed);
-        }
-        res
-    }
-}
-
-impl Serialize for SharedStoreMetric {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        serializer.serialize_u64(self.0.load(Ordering::Relaxed))
-    }
 }
 
 /// Reporter object which computes the process wall time and
@@ -680,72 +577,6 @@ impl SignalMetrics {
     }
 }
 
-/// Provides efficient way to record LatencyAggregateMetrics
-#[derive(Debug)]
-pub struct LatencyMetricsRecorder<'a> {
-    start_time: u64,
-    metric: &'a LatencyAggregateMetrics,
-}
-
-impl<'a> LatencyMetricsRecorder<'a> {
-    /// Const default construction.
-    fn new(metric: &'a LatencyAggregateMetrics) -> Self {
-        Self {
-            start_time: get_time_us(ClockType::Monotonic),
-            metric,
-        }
-    }
-}
-impl Drop for LatencyMetricsRecorder<'_> {
-    /// records aggregate (min/max/sum) for the given metric
-    /// This captures delta between self.start_time and current time
-    /// and updates min/max/sum metrics.
-    ///  self.start_time is recorded in new() and metrics are updated in drop
-    fn drop(&mut self) {
-        let delta_us = get_time_us(ClockType::Monotonic) - self.start_time;
-        self.metric.sum_us.add(delta_us);
-        let min_us = self.metric.min_us.fetch();
-        let max_us = self.metric.max_us.fetch();
-        if (0 == min_us) || (min_us > delta_us) {
-            self.metric.min_us.store(delta_us);
-        }
-        if (0 == max_us) || (max_us < delta_us) {
-            self.metric.max_us.store(delta_us);
-        }
-    }
-}
-
-/// Used to record Aggregate (min/max/sum) of latency metrics
-#[derive(Debug, Default, Serialize)]
-pub struct LatencyAggregateMetrics {
-    /// represents minimum value of the metrics in microseconds
-    pub min_us: SharedStoreMetric,
-    /// represents maximum value of the metrics in microseconds
-    pub max_us: SharedStoreMetric,
-    /// represents sum of the metrics in microseconds
-    pub sum_us: SharedIncMetric,
-}
-impl LatencyAggregateMetrics {
-    /// Const default construction.
-    pub const fn new() -> Self {
-        Self {
-            min_us: SharedStoreMetric::new(),
-            max_us: SharedStoreMetric::new(),
-            sum_us: SharedIncMetric::new(),
-        }
-    }
-
-    /// returns a latency recorder which captures stores start_time
-    /// and updates the actual metrics at the end of recorders lifetime.
-    /// in short instead of below 2 lines :
-    /// 1st for start_time_us = get_time_us()
-    /// 2nd for delta_time_us = get_time_us() - start_time; and metrics.store(delta_time_us)
-    /// we have just `_m = metrics.record_latency_metrics()`
-    pub fn record_latency_metrics(&self) -> LatencyMetricsRecorder<'_> {
-        LatencyMetricsRecorder::new(self)
-    }
-}
-
 /// Structure provides Metrics specific to VCPUs' mode of functioning.
 /// Sample_count or number of kvm exits for IO and MMIO VM exits are covered by:
 /// `exit_io_in`, `exit_io_out`, `exit_mmio_read` and , `exit_mmio_write`.
@@ -966,12 +797,13 @@ impl FirecrackerMetrics {
 mod tests {
     use std::io::{ErrorKind, LineWriter};
     use std::sync::Arc;
-    use std::sync::atomic::fence;
+    use std::sync::atomic::{Ordering, fence};
     use std::thread;
 
     use vmm_sys_util::tempfile::TempFile;
 
     use super::*;
+    use crate::metrics::IncMetric;
 
     #[test]
     fn test_init() {
