+++ src/vmm/src/mmds/mod.rs
@@ -18,7 +18,8 @@ use micro_http::{
 };
 use serde_json::{Map, Value};
 
-use crate::logger::{IncMetric, METRICS};
+use crate::logger::METRICS;
+use crate::metrics::IncMetric;
 use crate::mmds::data_store::{Mmds, MmdsDatastoreError as MmdsError, MmdsVersion, OutputFormat};
 use crate::mmds::token::PATH_TO_TOKEN;
 use crate::mmds::token_headers::{
